new kind of lambda parameter estimation and
precision is changed to double

+++ Gpufit/estimators/estimators.cuh
@@ -6,11 +6,11 @@
 
 __device__ void calculate_chi_square(
     int const estimator_id,
-    volatile float * chi_square,
+    volatile double * chi_square,
     int const point_index,
-    float const * data,
-    float const * value,
-    float const * weight,
+    double const * data,
+    double const * value,
+    double const * weight,
     int * state,
     char * user_info,
     std::size_t const user_info_size)
@@ -30,13 +30,13 @@ __device__ void calculate_chi_square(
 
 __device__ void calculate_gradient(
     int const estimator_id,
-    volatile float * gradient,
+    volatile double * gradient,
     int const point_index,
     int const parameter_index,
-    float const * data,
-    float const * value,
-    float const * derivative,
-    float const * weight,
+    double const * data,
+    double const * value,
+    double const * derivative,
+    double const * weight,
     char * user_info,
     std::size_t const user_info_size)
 {
@@ -59,10 +59,10 @@ __device__ void calculate_hessian(
     int const point_index,
     int const parameter_index_i,
     int const parameter_index_j,
-    float const * data,
-    float const * value,
-    float const * derivative,
-    float const * weight,
+    double const * data,
+    double const * value,
+    double const * derivative,
+    double const * weight,
     char * user_info,
     std::size_t const user_info_size)
 {
