+++ Gpufit/models/brown_dennis.cuh
@@ -41,11 +41,11 @@
 */
 
 __device__ void calculate_brown_dennis(
-    float const * parameters,
+    double const * parameters,
     int const n_fits,
     int const n_points,
-    float * value,
-    float * derivative,
+    double * value,
+    double * derivative,
     int const point_index,
     int const fit_index,
     int const chunk_index,
@@ -54,22 +54,22 @@ __device__ void calculate_brown_dennis(
 {
     // indices
 
-    float const x = (float)point_index / 5.f;
+    double const x = (double)point_index / 5.f;
 
     // parameters
 
-    float const * p = parameters;
+    double const * p = parameters;
     
     // value
 
-    float const arg1 = p[0] + p[1] * x - expf(x);
-    float const arg2 = p[2] + p[3] * sinf(x) - cosf(x);
+    double const arg1 = p[0] + p[1] * x - expf(x);
+    double const arg2 = p[2] + p[3] * sinf(x) - cosf(x);
 
     value[point_index] = arg1*arg1 + arg2*arg2;
 
     // derivative
 
-    float * current_derivative = derivative + point_index;
+    double * current_derivative = derivative + point_index;
 
     current_derivative[0 * n_points] = 2.f * arg1;
     current_derivative[1 * n_points] = 2.f * x * arg1;
